backend/requirements.txt
python-dotenv
langchain 
langchain-community 
langchain-openai
langchain-mongodb
pymongo
pydantic
chardet
fastapi
uvicorn